backend/requirements.txt
streamlit>=1.30.0
PyPDF2>=3.0.0
python-docx>=1.0.0
pdfminer.six>=20221030
jobspy>=0.7.0
plotly>=5.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
